// src/apis/models/FireBytes_Unified_Wave_ControlPanelTreeNodes_Dtos_ShowTablePanelInfoDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_ControlPanelTreeNodes_Dtos_ShowTablePanelInfoDto = {
    id?: string;
    displayName?: string | null;
    type?: string | null;
    graphicPanelId?: string;
    graphicPanel?: string | null;
    graphicType?: string | null;
    bladeVisualId?: string | null;
    graphicUrl?: string | null;
    graphicStoragePath?: string | null;
    modelId?: string;
    decompressingFileId?: string | null;
};